Stop caching `_drv`, `_root`, `_tail_cached`, and stop normalizing paths.
